refactor stepping test to extract common logic for other debug information test
